customer account jquery removed
